smallupdate

+++ HealthCareApp/HealthCareApp/ViewModels/AddMealViewModel.cs
@@ -220,7 +220,7 @@ namespace HealthCareApp.ViewModels
                 parameter.Calories.Focus();
                 return false;
             }
-            if (!int.TryParse(CaloriesVM, out _))
+            if (!double.TryParse(CaloriesVM, out _))
             {
                 MessageBox.Show("Please enter the calories in numeric format!", "Notification", MessageBoxButton.OK, MessageBoxImage.Warning);
                 parameter.Calories.Focus();
